requirements.txt
amazoncaptcha==0.5.11
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.4.2
outcome==1.3.0.post0
packaging==24.1
Pillow==9.5.0
playwright==1.48.0
playwright-stealth==1.0.6
prompt_toolkit==3.0.48
pycparser==2.22
pyee==12.0.0
PyJWT==2.9.0
PySocks==1.7.1
pytesseract==0.3.13
python-crontab==3.2.0
python-dateutil==2.9.0.post0
RapidFuzz==3.10.1
requests==2.30.0
selenium==4.25.0
setuptools==75.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
